feat: Add ProfileProvider to App and Home components
The code changes in the `App.jsx` and `Home.jsx` files add the `ProfileProvider` component to wrap the `Navigation` and `Home` components. This allows access to the user profile data throughout the application. The `ProfileProvider` is also added to the `ProtectedRoute` component in the `App.jsx` file. This change is necessary to provide the user profile data to the `Home` component and other components that require it.

Based on the recent repository commits, the commit message follows the convention of starting with a type (`feat` for new features) and providing a concise description of the changes made.