refactor: Add USER_KEY constant and update Navigation component
The code changes in the `constants.js` file add a new constant `USER_KEY` and update the `Navigation` component to use it. This constant is used to store the key for accessing the user data in local storage. The `Navigation` component is also updated to retrieve the user data from local storage and set the current user and role state variables accordingly. This change is necessary to improve the functionality of the navigation menu and ensure the correct user data is displayed.