refactor: Bulk create players and create user for each player
The code changes in the `apps.py` file add a new signal receiver function `load_initial_player_data` that is triggered after migrations are applied. This function bulk creates player objects using data from the `initial_players.csv` file. After creating the players, a user is manually created for each player to ensure that the `post_save` signal is triggered. This change improves the efficiency of creating players and ensures that the necessary user objects are created.