refactor: Add UserColorSquare component for displaying user color and name
The code changes in the `UserColorSquare.jsx` file add a new component called `UserColorSquare`. This component is responsible for displaying a colored square and the username. The color of the square is determined based on the `colorId` prop, which is mapped to a color value in the `colorDict` object. This change enhances the visual representation of users by showing their assigned colors alongside their names.