refactor: Update PlayerSerializer to include user field
The code changes in the `serializers.py` file update the `PlayerSerializer` to include the `user` field. This field was missing from the serializer and caused issues when serializing player objects. Adding the `user` field ensures that the serialized output includes the user information for each player. This change improves the accuracy and completeness of the serialized data.